+++ CHANGELOG.md
@@ -5,10 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.27.0] - 2024-12-11
+
+### Added
+
+### Removed
+
+### Changed
+
+### Fixed
+- Match framerate in `DefaultVideoFrameProcessorPipeline` with input stream
+
 ## [3.26.0] - 2024-10-07
 
 ### Added
 - Add interface for custom CDN paths for VideoFxProcessor
+
+- Allow setting framerate on `DefaultVideoTransformDevice`
 
 ### Removed